Move trace management to its own class
