+++ core/src/main/java/org/jruby/Ruby.java
@@ -40,7 +40,6 @@
 
 package org.jruby;
 
-import com.headius.invokebinder.Binder;
 import org.jcodings.specific.UTF8Encoding;
 import org.jruby.anno.FrameField;
 import org.jruby.anno.TypePopulator;
@@ -65,7 +64,6 @@ import org.jruby.ext.thread.SizedQueue;
 import org.jruby.ir.IRScope;
 import org.jruby.ir.IRScriptBody;
 import org.jruby.ir.runtime.IRReturnJump;
-import org.jruby.ir.runtime.IRRuntimeHelpers;
 import org.jruby.javasupport.Java;
 import org.jruby.javasupport.JavaClass;
 import org.jruby.javasupport.JavaPackage;
@@ -76,6 +74,7 @@ import org.jruby.management.InlineStats;
 import org.jruby.parser.StaticScope;
 import org.jruby.runtime.JavaSites;
 import org.jruby.runtime.MethodIndex;
+import org.jruby.runtime.TraceEventManager;
 import org.jruby.runtime.invokedynamic.InvokeDynamicSupport;
 import org.jruby.util.CommonByteLists;
 import org.jruby.util.JavaNameMangler;
@@ -186,14 +185,11 @@ import java.io.InputStream;
 import java.io.PrintStream;
 import java.io.PrintWriter;
 import java.lang.invoke.MethodHandle;
-import java.lang.invoke.MethodHandles;
-import java.lang.invoke.MutableCallSite;
 import java.lang.ref.WeakReference;
 import java.net.BindException;
 import java.nio.charset.Charset;
 import java.security.SecureRandom;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.EnumMap;
 import java.util.EnumSet;
@@ -218,7 +214,6 @@ import java.util.concurrent.atomic.AtomicLong;
 import java.util.function.Consumer;
 import java.util.function.ToIntFunction;
 import java.util.regex.Pattern;
-import java.util.stream.Collectors;
 
 import static java.lang.invoke.MethodHandles.explicitCastArguments;
 import static java.lang.invoke.MethodHandles.insertArguments;
@@ -250,21 +245,6 @@ public final class Ruby implements Constantizable {
      * The logger used to log relevant bits.
      */
     private static final Logger LOG = LoggerFactory.getLogger(Ruby.class);
-    private static final MethodHandles.Lookup LOOKUP = MethodHandles.lookup();
-    public static final MethodHandle TRACE_OFF = Binder
-            .from(void.class, ThreadContext.class, IRubyObject.class, RubyEvent.class, String.class, String.class, int.class)
-            .drop(1, 5)
-            .identity();
-    public static final MethodHandle TRACE_ON = Binder
-            .from(void.class, ThreadContext.class, IRubyObject.class, RubyEvent.class, String.class, String.class, int.class)
-            .invokeStaticQuiet(LOOKUP, IRRuntimeHelpers.class, "callTrace");
-    public static final MethodHandle B_TRACE_OFF = Binder
-            .from(void.class, ThreadContext.class, Block.class, RubyEvent.class, String.class, String.class, int.class)
-            .drop(1, 5)
-            .identity();
-    public static final MethodHandle B_TRACE_ON = Binder
-            .from(void.class, ThreadContext.class, Block.class, RubyEvent.class, String.class, String.class, int.class)
-            .invokeStaticQuiet(LOOKUP, IRRuntimeHelpers.class, "callTrace");
 
     /**
      * Create and initialize a new JRuby runtime. The properties of the
@@ -3121,201 +3101,11 @@ public final class Ruby implements Constantizable {
         return boundMethods;
     }
 
-    private static final EnumSet<RubyEvent> interest =
-            EnumSet.of(
-                    RubyEvent.C_CALL,
-                    RubyEvent.C_RETURN,
-                    RubyEvent.CALL,
-                    RubyEvent.CLASS,
-                    RubyEvent.END,
-                    RubyEvent.LINE,
-                    RubyEvent.RAISE,
-                    RubyEvent.RETURN
-            );
+    private final TraceEventManager traceEvents = new TraceEventManager(this);
 
-    public MutableCallSite getCallTrace() {
-        return callTrace;
+    public TraceEventManager getTraceEvents() {
+        return traceEvents;
     }
-
-    public MutableCallSite getBcallTrace() {
-        return bcallTrace;
-    }
-
-    public static class CallTraceFuncHook extends EventHook {
-        private RubyProc traceFunc;
-        private final ThreadContext thread; // if non-null only call traceFunc if it is from this thread.
-
-        public CallTraceFuncHook(ThreadContext context) {
-            this.thread = context;
-        }
-
-        public void setTraceFunc(RubyProc traceFunc) {
-            this.traceFunc = traceFunc;
-        }
-
-        public void eventHandler(ThreadContext context, String eventName, String file, int line, String name, IRubyObject type) {
-            if (context.isWithinTrace()) return;
-            if (thread != null && thread != context) return;
-
-            if (file == null) file = "(ruby)";
-            if (type == null) type = context.nil;
-
-            Ruby runtime = context.runtime;
-            RubyBinding binding = RubyBinding.newBinding(runtime, context.currentBinding());
-
-            // FIXME: Ultimately we should be getting proper string for this event type
-            switch(eventName) {
-                case "c_return":
-                    eventName = "c-return";
-                    break;
-                case "c_call":
-                    eventName = "c-call";
-                    break;
-            };
-
-            context.preTrace();
-            try {
-                traceFunc.call(context, new IRubyObject[]{
-                        runtime.newString(eventName), // event name
-                        runtime.newString(file), // filename
-                        runtime.newFixnum(line), // line numbers should be 1-based
-                        name != null ? runtime.newSymbol(name) : runtime.getNil(),
-                        binding,
-                        type
-                });
-            } finally {
-                context.postTrace();
-            }
-        }
-
-        @Override
-        public boolean equals(Object other) {
-            if (!(other instanceof CallTraceFuncHook)) return false;
-
-            return super.equals(other) && thread == ((CallTraceFuncHook) other).thread;
-        }
-
-        @Override
-        public boolean isInterestedInEvent(RubyEvent event) {
-            return interest.contains(event);
-        }
-
-        public ThreadContext getThread() {
-            return thread;
-        }
-
-        @Override
-        public EnumSet<RubyEvent> eventSet() {
-            return interest;
-        }
-    };
-
-    private final CallTraceFuncHook callTraceFuncHook = new CallTraceFuncHook(null);
-
-    public synchronized void addEventHook(EventHook hook) {
-        if (!RubyInstanceConfig.FULL_TRACE_ENABLED && hook.needsDebug()) {
-            // without full tracing, many events will not fire
-            getWarnings().warn("tracing (e.g. set_trace_func) will not capture all events without --debug flag");
-        }
-
-        EventHook[] hooks = eventHooks;
-        EventHook[] newHooks = Arrays.copyOf(hooks, hooks.length + 1);
-        newHooks[hooks.length] = hook;
-        eventHooks = newHooks;
-        callTrace.setTarget(TRACE_ON);
-        bcallTrace.setTarget(B_TRACE_ON);
-        hasEventHooks = true;
-    }
-
-    public synchronized void removeEventHook(EventHook hook) {
-        EventHook[] hooks = eventHooks;
-
-        if (hooks.length == 0) return;
-
-        int pivot = -1;
-        for (int i = 0; i < hooks.length; i++) {
-            if (hooks[i].equals(hook)) {
-                pivot = i;
-                break;
-            }
-        }
-
-        if (pivot == -1) return; // No such hook found.
-
-        EventHook[] newHooks = new EventHook[hooks.length - 1];
-        // copy before and after pivot into the new array but don't bother
-        // to arraycopy if pivot is first/last element of the old list.
-        if (pivot != 0) System.arraycopy(hooks, 0, newHooks, 0, pivot);
-        if (pivot != hooks.length-1) System.arraycopy(hooks, pivot + 1, newHooks, pivot, hooks.length - (pivot + 1));
-
-        eventHooks = newHooks;
-        hasEventHooks = newHooks.length > 0;
-        if (newHooks.length == 0) callTrace.setTarget(TRACE_OFF);
-        if (newHooks.length == 0) bcallTrace.setTarget(B_TRACE_OFF);
-    }
-
-    public void setTraceFunction(RubyProc traceFunction) {
-        setTraceFunction(callTraceFuncHook, traceFunction);
-    }
-
-    public void setTraceFunction(CallTraceFuncHook hook, RubyProc traceFunction) {
-        removeEventHook(hook);
-
-        if (traceFunction == null) return;
-
-        hook.setTraceFunc(traceFunction);
-        addEventHook(hook);
-    }
-
-    /**
-     * Remove all event hooks which are associated with a particular thread.
-     * @param context the context of the ruby thread we are interested in.
-     */
-    public void removeAllCallEventHooksFor(ThreadContext context) {
-        if (eventHooks.length == 0) return;
-
-        List<EventHook> hooks = new ArrayList<>(Arrays.asList(eventHooks));
-
-        hooks = hooks.stream().filter(hook ->
-                !(hook instanceof CallTraceFuncHook) || !((CallTraceFuncHook) hook).getThread().equals(context)
-        ).collect(Collectors.toList());
-
-        EventHook[] newHooks = new EventHook[hooks.size()];
-        eventHooks = hooks.toArray(newHooks);
-        hasEventHooks = hooks.size() > 0;
-    }
-
-    public void callEventHooks(ThreadContext context, RubyEvent event, String file, int line, String name, IRubyObject type) {
-        if (context.isEventHooksEnabled()) {
-            EventHook hooks[] = eventHooks;
-
-            for (EventHook eventHook: hooks) {
-                if (eventHook.isInterestedInEvent(event)) {
-                    IRubyObject klass = context.nil;
-                    if (type instanceof RubyModule) {
-                        if (((RubyModule) type).isIncluded()) {
-                            klass = ((RubyModule) type).getOrigin();
-                        } else if (((RubyModule) type).isSingleton()) {
-                            klass = ((MetaClass) type).getAttached();
-                        }
-                    }
-                    eventHook.event(context, event, file, line, name, klass);
-                }
-            }
-        }
-    }
-
-    public boolean hasEventHooks() {
-        return hasEventHooks;
-    }
-
-    private final MutableCallSite callTrace = new MutableCallSite(
-            TRACE_OFF
-    );
-
-    private final MutableCallSite bcallTrace = new MutableCallSite(
-            B_TRACE_OFF
-    );
 
     public GlobalVariables getGlobalVariables() {
         return globalVariables;
@@ -5452,10 +5242,6 @@ public final class Ruby implements Constantizable {
 
     private final RubySymbol.SymbolTable symbolTable = new RubySymbol.SymbolTable(this);
 
-    private static final EventHook[] EMPTY_HOOKS = new EventHook[0];
-    private volatile EventHook[] eventHooks = EMPTY_HOOKS;
-    private boolean hasEventHooks;
-
     private boolean abortOnException = false; // Thread.abort_on_exception
     private boolean reportOnException = true; // Thread.report_on_exception
     private boolean doNotReverseLookupEnabled = false;
@@ -5984,6 +5770,41 @@ public final class Ruby implements Constantizable {
     @Deprecated
     public RaiseException newFrozenError(String objectType, boolean runtimeError) {
         return newRaiseException(getFrozenError(), str(this, "can't modify frozen ", ids(this, objectType)));
+    }
+
+    @Deprecated
+    public synchronized void addEventHook(EventHook hook) {
+        traceEvents.addEventHook(hook);
+    }
+
+    @Deprecated
+    public synchronized void removeEventHook(EventHook hook) {
+        traceEvents.removeEventHook(hook);
+    }
+
+    @Deprecated
+    public void setTraceFunction(RubyProc traceFunction) {
+        traceEvents.setTraceFunction(traceFunction);
+    }
+
+    @Deprecated
+    public void setTraceFunction(TraceEventManager.CallTraceFuncHook hook, RubyProc traceFunction) {
+        traceEvents.setTraceFunction(hook, traceFunction);
+    }
+
+    @Deprecated
+    public void removeAllCallEventHooksFor(ThreadContext context) {
+        traceEvents.removeAllCallEventHooksFor(context);
+    }
+
+    @Deprecated
+    public void callEventHooks(ThreadContext context, RubyEvent event, String file, int line, String name, IRubyObject type) {
+        traceEvents.callEventHooks(context, event, file, line, name, type);
+    }
+
+    @Deprecated
+    public boolean hasEventHooks() {
+        return traceEvents.hasEventHooks();
     }
 
 }
