+++ core/src/main/java/org/jruby/ir/runtime/IRRuntimeHelpers.java
@@ -69,6 +69,7 @@ import org.jruby.runtime.IRBlockBody;
 import org.jruby.runtime.JavaSites.IRRuntimeHelpersSites;
 import org.jruby.runtime.RubyEvent;
 import org.jruby.runtime.ThreadContext;
+import org.jruby.runtime.TraceEventManager;
 import org.jruby.runtime.Visibility;
 import org.jruby.runtime.backtrace.RubyStackTraceElement;
 import org.jruby.runtime.builtin.IRubyObject;
@@ -2451,18 +2452,18 @@ public class IRRuntimeHelpers {
 
     @JIT
     public static void callTrace(ThreadContext context, IRubyObject selfClass, RubyEvent event, String name, String filename, int line) {
-        Ruby runtime = context.runtime;
-        if (runtime.hasEventHooks()) {
+        TraceEventManager traceEvents = context.traceEvents;
+        if (traceEvents.hasEventHooks()) {
             // FIXME: Try and statically generate END linenumber instead of hacking it.
             int linenumber = line == -1 ? context.getLine() : line;
 
-            runtime.callEventHooks(context, event, filename, linenumber, name, selfClass);
+            traceEvents.callEventHooks(context, event, filename, linenumber, name, selfClass);
         }
     }
 
-    public static void traceRaise(ThreadContext context, RubyException exception) {
-        Ruby runtime = context.runtime;
-        if (runtime.hasEventHooks()) {
+    public static void traceRaise(ThreadContext context) {
+        TraceEventManager traceEvents = context.traceEvents;
+        if (traceEvents.hasEventHooks()) {
             RubyStackTraceElement backtraceElement = context.getSingleBacktrace();
             String file = backtraceElement.getFileName();
             int line = backtraceElement.getLineNumber();
@@ -2470,18 +2471,18 @@ public class IRRuntimeHelpers {
             // FIXME: Try and statically generate END linenumber instead of hacking it.
             int linenumber = line == -1 ? context.getLine() : line;
 
-            runtime.callEventHooks(context, RubyEvent.RAISE, file, linenumber, null, context.nil);
+            traceEvents.callEventHooks(context, RubyEvent.RAISE, file, linenumber, null, context.nil);
         }
     }
 
     @JIT
     public static void callTrace(ThreadContext context, Block selfBlock, RubyEvent event, String name, String filename, int line) {
-        Ruby runtime = context.runtime;
-        if (runtime.hasEventHooks()) {
+        TraceEventManager traceEvents = context.traceEvents;
+        if (traceEvents.hasEventHooks()) {
             // FIXME: Try and statically generate END linenumber instead of hacking it.
             int linenumber = line == -1 ? context.getLine() : line;
 
-            runtime.callEventHooks(context, event, filename, linenumber, name, selfBlock.getFrameClass());
+            traceEvents.callEventHooks(context, event, filename, linenumber, name, selfBlock.getFrameClass());
         }
     }
 
@@ -2490,7 +2491,7 @@ public class IRRuntimeHelpers {
         // FIXME: Try and statically generate END linenumber instead of hacking it.
         int linenumber = line == -1 ? context.getLine() : line;
 
-        context.runtime.callEventHooks(context, event, filename, linenumber, name, selfBlock.getFrameClass());
+        context.traceEvents.callEventHooks(context, event, filename, linenumber, name, selfBlock.getFrameClass());
     }
 
     public static void warnSetConstInRefinement(ThreadContext context, IRubyObject self) {
